+++ src/test/java/jscover/MainParsingTest.java
@@ -362,6 +362,7 @@ public class MainParsingTest {
         assertThat(main.isFileSystem(), equalTo(false));
         assertThat(main.showCharSets(), equalTo(false));
         assertThat(main.isRegExpTest(), equalTo(false));
+        assertThat(main.isGenerateFiles(), equalTo(false));
     }
 
     @Test
@@ -407,6 +408,12 @@ public class MainParsingTest {
     public void shouldParseEncodingHelp() {
         assertThat(main.parse(new String[]{"-ws","encoding"}).showCharSets(), equalTo(false));
         assertThat(main.parse(new String[]{"-h","encoding"}).showCharSets(), equalTo(true));
+        assertThat(main.getExitStatus(), equalTo(0));
+    }
+
+    @Test
+    public void shouldParseGenerateFilesTest() {
+        assertThat(main.parse(new String[]{"-gf"}).isGenerateFiles(), equalTo(true));
         assertThat(main.getExitStatus(), equalTo(0));
     }